Backend: Changing functions cart, favorite. Testing functions.

+++ backend/src/favorites/favorites.service.ts
@@ -1,10 +1,43 @@
 import { ForbiddenException, Injectable } from "@nestjs/common";
 import { PrismaService } from "../utils/prisma.service";
-import { Catalog } from "../models/Catalog";
+import { Catalog, shortCatalog } from "../models/Catalog";
+import { Favorite } from "../models/Favorite";
 
 @Injectable()
 export class FavoritesService {
     constructor(private prismaService: PrismaService) {}
+
+    /**
+     * Получение избранных товаров.
+     * @param user_id - id пользователя
+     */
+    async getFavorites(user_id: number) {
+        const favorites: Favorite[] =
+            await this.prismaService.favorite.findMany({
+                where: {
+                    user_id,
+                },
+                include: {
+                    catalogId: true,
+                },
+            });
+
+        if (!favorites) {
+            throw new ForbiddenException("Товар не найден");
+        }
+
+        const shortFavorites: shortCatalog[] = favorites.map((favorite) => {
+            return {
+                id: favorite.catalogId.id,
+                title: favorite.catalogId.title,
+                image: favorite.catalogId.images[0],
+                cost: favorite.catalogId.cost,
+                status: favorite.catalogId.status,
+            };
+        });
+
+        return shortFavorites;
+    }
 
     /**
      * Добавление товара в избранное.
@@ -36,11 +69,24 @@ export class FavoritesService {
      * @param catalog_id - id товара
      */
     async removeFavorite(user_id: number, catalog_id: number) {
-        return this.prismaService.favorite.deleteMany({
+        const _favorite = await this.prismaService.favorite.findFirst({
             where: {
                 user_id,
                 catalog_id,
             },
         });
+
+        if (!_favorite) throw new ForbiddenException("Товар не найден");
+
+        await this.prismaService.favorite.deleteMany({
+            where: {
+                user_id,
+                catalog_id,
+            },
+        });
+
+        return {
+            status: "Удалено",
+        };
     }
 }
